Behaviour Change.
The cutoffDate property has been removed from the Accommodation class. Instead, the cutoff date for a booking can only be accessed at the booking level via the accommCutOffDate property.

In addition, if a non-null accomCutOffDate value is not provided on a call to AddOrUpdateVenueBookingAsync, the cutoff date will be cleared from the booking. This is a behaviour modification, because previously, specifying no value would mean that no update was made.